djangocrud/requirements.txt
#django==2.2.12
django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
#djongo==1.3.2